requirements.txt
pandas
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
